+++ neo/idlib/Parser.cpp
@@ -1511,7 +1511,7 @@ typedef struct operator_s
 
 typedef struct value_s
 {
-	signed long int intvalue;
+	signed int intvalue; // DG: use int instead of long for 64bit compatibility
 	double floatvalue;
 	int parentheses;
 	struct value_s* prev, *next;
@@ -1606,7 +1606,7 @@ int PC_OperatorPriority( int op )
 
 #define FreeOperator(op)
 
-int idParser::EvaluateTokens( idToken* tokens, signed long int* intvalue, double* floatvalue, int integer )
+int idParser::EvaluateTokens( idToken* tokens, signed int* intvalue, double* floatvalue, int integer )
 {
 	operator_t* o, *firstoperator, *lastoperator;
 	value_t* v, *firstvalue, *lastvalue, *v1, *v2;
@@ -2117,7 +2117,7 @@ int idParser::EvaluateTokens( idToken* tokens, signed long int* intvalue, double
 idParser::Evaluate
 ================
 */
-int idParser::Evaluate( signed long int* intvalue, double* floatvalue, int integer )
+int idParser::Evaluate( signed int* intvalue, double* floatvalue, int integer )
 {
 	idToken token, *firsttoken, *lasttoken;
 	idToken* t, *nexttoken;
@@ -2224,7 +2224,7 @@ int idParser::Evaluate( signed long int* intvalue, double* floatvalue, int integ
 idParser::DollarEvaluate
 ================
 */
-int idParser::DollarEvaluate( signed long int* intvalue, double* floatvalue, int integer )
+int idParser::DollarEvaluate( signed int* intvalue, double* floatvalue, int integer )
 {
 	int indent, defined = false;
 	idToken token, *firsttoken, *lasttoken;
@@ -2345,7 +2345,7 @@ idParser::Directive_elif
 */
 int idParser::Directive_elif()
 {
-	signed long int value;
+	signed int value; // DG: use int instead of long for 64bit compatibility
 	int type, skip;
 	
 	idParser::PopIndent( &type, &skip );
@@ -2370,7 +2370,7 @@ idParser::Directive_if
 */
 int idParser::Directive_if()
 {
-	signed long int value;
+	signed int value; // DG: use int instead of long for 64bit compatibility
 	int skip;
 	
 	if( !idParser::Evaluate( &value, NULL, true ) )
@@ -2477,7 +2477,7 @@ idParser::Directive_eval
 */
 int idParser::Directive_eval()
 {
-	signed long int value;
+	signed int value; // DG: use int instead of long for 64bit compatibility
 	idToken token;
 	char buf[128];
 	
@@ -2644,7 +2644,7 @@ idParser::DollarDirective_evalint
 */
 int idParser::DollarDirective_evalint()
 {
-	signed long int value;
+	signed int value; // DG: use int instead of long for 64bit compatibility
 	idToken token;
 	char buf[128];
 	
@@ -2697,7 +2697,7 @@ int idParser::DollarDirective_evalfloat()
 	token = buf;
 	token.type = TT_NUMBER;
 	token.subtype = TT_FLOAT | TT_LONG | TT_DECIMAL | TT_VALUESVALID;
-	token.intvalue = ( unsigned long ) fabs( value );
+	token.intvalue = ( unsigned int ) fabs( value ); // DG: use int instead of long for 64bit compatibility
 	token.floatvalue = fabs( value );
 	idParser::UnreadSourceToken( &token );
 	if( value < 0 )
